issues 6, 8, 9

See merge request j2mascar/mobileapplications!27

+++ composeApp/src/commonMain/kotlin/com/example/foodmanager/domain/model/ShoppingItem.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ShoppingItem(
-    val id: Int? = null, // Database auto-generates this as a numeric ID
+    val id: Long? = null, // Database auto-generates this as a numeric ID
     val shopping_list_id: String, // Must be a String for the UUID
     val name: String,
     val amount: Double,
